refactor: restructure assets folder
