+++ frontend/src/assets/homepage/map-image.png
