+++ frontend/src/assets/icons/ability-avatar.png
