+++ frontend/src/assets/icons/builder-avatar.png
