+++ frontend/src/assets/icons/cancel-icon.png
