+++ frontend/src/assets/icons/fortified.png
