+++ frontend/src/assets/icons/gold.png
