+++ frontend/src/assets/icons/king-attack.png
