+++ frontend/src/assets/icons/king-regen.png
