+++ frontend/src/assets/icons/light.png
