+++ frontend/src/assets/icons/magic.png
