+++ frontend/src/assets/icons/normal.png
