+++ frontend/src/assets/icons/unarmored.png
