+++ frontend/src/assets/icons/unit-avatar.png
