+++ frontend/src/shared/Services/get-icons.tsx
@@ -1,22 +1,22 @@
-import cancelIcon from "../../assets/cancel-icon.png";
-import chaosAttack from "../../assets/chaos.png";
-import fortifiedDefense from "../../assets/fortified.png";
-import heavyDefense from "../../assets/heavy.png";
-import lightDefense from "../../assets/light.png";
-import magicAttack from "../../assets/magic.png";
-import mediumDefense from "../../assets/medium.png";
-import normalAttack from "../../assets/normal.png";
-import pierceAttack from "../../assets/pierce.png";
-import siegeAttack from "../../assets/siege.png";
-import tierFive from "../../assets/tier-five.png";
-import tierFour from "../../assets/tier-four.png";
-import tierOne from "../../assets/tier-one.png";
-import tierSix from "../../assets/tier-six.png";
-import tierThree from "../../assets/tier-three.png";
-import tierTwo from "../../assets/tier-two.png";
-import unarmoredDefense from "../../assets/unarmored.png";
-import unitAvatar from "../../assets/unit-avatar.png";
-import unknownIcon from "../../assets/unknown-icon.png";
+import cancelIcon from "../../assets/icons/cancel-icon.png";
+import chaosAttack from "../../assets/icons/chaos.png";
+import fortifiedDefense from "../../assets/icons/fortified.png";
+import heavyDefense from "../../assets/icons/heavy.png";
+import lightDefense from "../../assets/icons/light.png";
+import magicAttack from "../../assets/icons/magic.png";
+import mediumDefense from "../../assets/icons/medium.png";
+import normalAttack from "../../assets/icons/normal.png";
+import pierceAttack from "../../assets/icons/pierce.png";
+import siegeAttack from "../../assets/icons/siege.png";
+import tierFive from "../../assets/icons/tier-five.png";
+import tierFour from "../../assets/icons/tier-four.png";
+import tierOne from "../../assets/icons/tier-one.png";
+import tierSix from "../../assets/icons/tier-six.png";
+import tierThree from "../../assets/icons/tier-three.png";
+import tierTwo from "../../assets/icons/tier-two.png";
+import unarmoredDefense from "../../assets/icons/unarmored.png";
+import unitAvatar from "../../assets/icons/unit-avatar.png";
+import unknownIcon from "../../assets/icons/unknown-icon.png";
 
 export const getAttackIcon = (attackType: string | null): string => {
   switch (attackType) {
